+++ src/main/java/org/folio/search/service/InstanceChildrenResourceService.java
@@ -1,5 +1,9 @@
 package org.folio.search.service;
 
+import static org.apache.commons.lang3.StringUtils.startsWith;
+import static org.folio.search.utils.SearchConverterUtils.getResourceSource;
+import static org.folio.search.utils.SearchUtils.SOURCE_CONSORTIUM_PREFIX;
+
 import java.util.List;
 import java.util.Map;
 import java.util.stream.Collectors;
@@ -9,6 +13,7 @@ import org.folio.search.domain.dto.ResourceEventType;
 import org.folio.search.model.types.ResourceType;
 import org.folio.search.service.consortium.ConsortiumTenantProvider;
 import org.folio.search.service.converter.preprocessor.extractor.ChildResourceExtractor;
+import org.folio.search.utils.SearchConverterUtils;
 import org.springframework.stereotype.Component;
 
 /**
@@ -34,8 +39,22 @@ public class InstanceChildrenResourceService {
     if (extractors == null) {
       return;
     }
+
     var shared = consortiumTenantProvider.isCentralTenant(tenantId);
-    extractors.forEach(resourceExtractor -> resourceExtractor.persistChildren(shared, events));
+    var noShadowCopiesInstanceEvents = events.stream()
+      .filter(resourceEvent -> {
+        if (ResourceType.INSTANCE.getName().equals(resourceEvent.getResourceName())) {
+          return !startsWith(getResourceSource(resourceEvent), SOURCE_CONSORTIUM_PREFIX);
+        }
+        return true;
+      })
+      .toList();
+    var eventsForResourceSharing = events.stream()
+      .filter(SearchConverterUtils::isUpdateEventForResourceSharing)
+      .toList();
+    extractors.forEach(resourceExtractor -> resourceExtractor.persistChildren(shared, noShadowCopiesInstanceEvents));
+    extractors.forEach(resourceExtractor ->
+      resourceExtractor.persistChildrenForResourceSharing(shared, eventsForResourceSharing));
   }
 
   public void persistChildrenOnReindex(String tenantId, ResourceType resourceType,
@@ -50,5 +69,4 @@ public class InstanceChildrenResourceService {
       .toList();
     persistChildren(tenantId, resourceType, events);
   }
-
 }