+++ src/main/java/org/folio/search/service/converter/preprocessor/extractor/ChildResourceExtractor.java
@@ -2,6 +2,9 @@ package org.folio.search.service.converter.preprocessor.extractor;
 
 import static java.util.Collections.emptyList;
 import static java.util.Collections.emptySet;
+import static java.util.stream.Collectors.groupingBy;
+import static java.util.stream.Collectors.mapping;
+import static java.util.stream.Collectors.toList;
 import static org.apache.commons.collections4.MapUtils.getObject;
 import static org.folio.search.utils.SearchConverterUtils.getNewAsMap;
 
@@ -13,6 +16,7 @@ import java.util.Map;
 import java.util.Objects;
 import java.util.Set;
 import lombok.RequiredArgsConstructor;
+import lombok.extern.log4j.Log4j2;
 import org.folio.search.domain.dto.ResourceEvent;
 import org.folio.search.domain.dto.ResourceEventType;
 import org.folio.search.model.entity.ChildResourceEntityBatch;
@@ -20,6 +24,7 @@ import org.folio.search.model.types.ResourceType;
 import org.folio.search.service.reindex.jdbc.InstanceChildResourceRepository;
 
 @RequiredArgsConstructor
+@Log4j2
 public abstract class ChildResourceExtractor {
 
   private final InstanceChildResourceRepository repository;
@@ -39,7 +44,7 @@ public abstract class ChildResourceExtractor {
       .map(ResourceEvent::getId)
       .toList();
     if (!instanceIdsForDeletion.isEmpty()) {
-      repository.deleteByInstanceIds(instanceIdsForDeletion);
+      repository.deleteByInstanceIds(instanceIdsForDeletion, null);
     }
 
     var eventsForSaving = events.stream()
@@ -57,6 +62,18 @@ public abstract class ChildResourceExtractor {
       entities.addAll(entitiesFromEvent);
     });
     repository.saveAll(new ChildResourceEntityBatch(new ArrayList<>(entities), relations));
+  }
+
+  public void persistChildrenForResourceSharing(boolean shared, List<ResourceEvent> events) {
+    var eventsForSharingByTenant = events.stream()
+      .collect(groupingBy(ResourceEvent::getTenant, mapping(ResourceEvent::getId, toList())));
+    eventsForSharingByTenant.forEach((tenant, instanceIds) -> {
+      if (Boolean.TRUE.equals(shared)) {
+        log.warn("Update event for instance sharing is supposed to be for member tenant,"
+          + " but received for central tenant: {}, eventId: {}", tenant, String.join(",", instanceIds));
+      }
+      repository.deleteByInstanceIds(instanceIds, tenant);
+    });
   }
 
   private List<Map<String, Object>> extractEntities(ResourceEvent event) {